+++ ebustoolbox/urls.py
@@ -13,6 +13,8 @@ urlpatterns = [
         views.long_running_task_status_view,
         name="long_running_task_status_view",
     ),
+    path("scenarios/", views.scenarios, name="scenarios"),
+    path("usergroups/", views.usergroups, name="usergroups"),
     path("result/", views.result_view, name="result"),
     path("", views.home_view, name="home"),
     path("generate_zip/<uuid:task_id>", views.generate_zip, name="generate_zip"),